Testing: Generalize integration test setup/teardown boilerplate code
